+++ src/App.css
@@ -1,5 +1,10 @@
 /* App.css */
-
+/*
+#page-container {
+  padding-top:3%;
+  padding-bottom:5%;
+}
+*/
 body {
   font-family: Arial, sans-serif;
   margin: 0;
@@ -22,7 +27,6 @@ text-align: center;
   color: white;
 }
 
-
 .App-link {
 color: #61dafb;
 }
@@ -34,12 +38,35 @@ nav {
 }
 
 .App nav{
-  position:fixed;
   width: 100%;
   height: 5%;
+}
+
+#Header{
+  position:fixed;
+  width:100%;
+  height:3%;
+  /*background-color:#ffd700;*/
+  background-color:rgba(0,100,100,.5);
+  color:#000080;
+  font-size:100%;
+  top:0;
+}
+
+#Footer {
+  position:fixed;
+  width:100%;
+  height:5%;
+  bottom:0;
+  /*background-color: #000080;*/
+  background-color:rgba(0,100,100,.5);
 }
 
 .navButton{
   color:blue;
   background-color: #ffd700;
+}
+
+#SearchBar {
+  display:inline-block;
 }
