--- src/package.json
@@ -1,19 +1,0 @@
-{
-    "name": "msu-news-app",
-    "version": "1.0.0",
-    "description": "News app for the MSU news department",
-    "main": "home.js",
-    "scripts": {
-      "start": "node your-entry-file.js"
-    },
-    "keywords": [],
-    "authors": "Noah Lemmon, Luke Felisky, Jason Kelley",
-    "license": "ISC",
-    "dependencies": {
-      "react": "^17.0.2",
-      "react-dom": "^17.0.2",
-      "react-router-dom": "^6.0.2",
-      "reactjs-popup": "^2.0.6"
-    }
-  }
-
